+++ src/bunit.core/RazorTesting/RazorTestBase.cs
@@ -7,7 +7,7 @@ namespace Bunit.RazorTesting
 	/// <summary>
 	/// Represents a component used to define tests in Razor files.
 	/// </summary>
-	public abstract class RazorTestBase : TestContextBase, ITestContext, IComponent
+	public abstract class RazorTestBase : TestContextBase, IComponent
 	{
 		/// <summary>
 		/// Gets the name of the test, which is displayed by the test runner/explorer.